Fix value in HTTP examples

+++ examples/Boards_With_HTTP_API/Arduino_GSM/Arduino_GSM.ino
@@ -138,7 +138,7 @@ void setup()
 void loop() {
   String response;
 
-  int value = millis();
+  unsigned long value = millis();
 
   // Send value to the cloud
   // similar to Blynk.virtualWrite()
